Adds the Button component along with oceanside colorscheme
Removes redundant styling
