+++ stories/Button.stories.tsx
@@ -1,13 +1,9 @@
 import { Story, Meta } from '@storybook/react'
-
-import { Button, ButtonProps } from './Button'
+import { Button, ButtonProps } from '../components/Button'
 
 export default {
-  title: 'Example/Button',
+  title: 'Component/Atoms/Button',
   component: Button,
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
 } as Meta
 
 const Template: Story<ButtonProps> = (args) => <Button {...args} />
@@ -15,22 +11,10 @@ const Template: Story<ButtonProps> = (args) => <Button {...args} />
 export const Primary = Template.bind({})
 Primary.args = {
   primary: true,
-  label: 'Button',
+  label: 'Manage Claim',
 }
 
 export const Secondary = Template.bind({})
 Secondary.args = {
-  label: 'Button',
-}
-
-export const Large = Template.bind({})
-Large.args = {
-  size: 'large',
-  label: 'Button',
-}
-
-export const Small = Template.bind({})
-Small.args = {
-  size: 'small',
-  label: 'Button',
+  label: 'Secondary Button',
 }
